Attempt to optimise vvorset for space a little
With a vclock per entry, assuming a reasonably static set of actors
and the same actors adding each entry, there is a lot of repeated
actor ids in the datastructure. Riak_Kv uses 8byte actor IDs.

Tests showed that for 100 entries, each with a number of actors
(concurrent adds) the size of the vvorset was not inconsequential.

Added an actor list, which is used to replace each 8byte actor ID
with a single int. These are encoded by erlang to 3 bytes.

Tests show a saving of 9Kb on an 100 entry set with high concurrency
for adds.