Change semantics of multi so a remove is also a `reset`
In the case of concurrent update |  remove the previous map
semantics where confusing, in that a concurrent update would
maintain the seen state at the updating replica, winning over the remove
but not keeping any unseen updates. This seems inconsistent.

Remove - reset means that when a field is removed from the replica
it first sets the CRDT at at Field to it's original state with an
update. For example a PNCounter is incremented to zero. A Set has
all it's active members removed. This way a concurrent update still
wins over a remove, in that the field remains, but only the state
that was not actively removed remains.