Merge tombstones in the Map
If two tombstoned replicas merge the tombstone is lost. A bug in the
map meant that only fields that were present in at least one replica
were persisted.

A side effect of merging tombstones and keeping them is that in some
edge cases a `reset` can be magnified.

Imagine replicas A, B, and C. A updates a counter at field F by 10
B updates a counter at field F by 10. A and B merge.
A removes the counter at F (resulting in actor A decremting by 20)
B removes the counter at F (resulting in actor B decremening by 20)
C inrements a counter at F by 10.
A, B and C merge. The counter at F is present and it's value is -10.

This makes sense as a remove is also a decrement of all seen values
but may confuse a user.

Sets are unaffected as their removes are idempotent.