Add a tombestone-less Set and Map
Based on http://arxiv.org/abs/1210.3368 (but stated based)
The Set and Map use the same design. Read the documentation on
riak_dt_orswot for details.

NOTE: This does not mean no GC ever. The size of the version vector
at the head of the Map and Set will grow without bound over time
due to Riak actor churn / fallbacks. But for riak2.0 I believe
this is stable enough.

Also added a G-Set for completeness.