Add comments around the merge function for common keys
The keys that are common to each set need no special merge, just
pick one, either side's element's dot will be contained in the merged
clock for the merged set. However, `equal/2' expects two merged sets
to be structually equal, regardless of merge order. For that reason
we take the trouble to always chose the highest dot.

We can revisit this if we think it is too ineffecient.