Address review comments
Remove the actor list optimisation as it confuses merge and is
only needed for to/from binary.

Store the list of dots against each entry, this means that merge
is just a merge of birth dots. Storing a single dot was a bug.

Also, on merge when an element is only present on one side, drop
all 'seen' dots from the dot list for the element. This is just an
optimisation.

Thanks again to Carlos Baquero who took the time to speak with me
about this stuff today.