A map with more fine grained causality->value mapping
Makes the end result of remove deterministic at the cost of
storing more state between merge and update.

Essentially stores siblings of merging replicas values for each
field against a birth date for that state. An update to the field
generates a new dot and merges to a single value.