Explore an interface for mapping between CRDTs.
What I'm looking at integrating is keys, that are derived from each
other in Riak KV; for instance, I could have a counter key, and
specified in the metadata, have any updates to it propogate to another
key which is a boolean key based on a particular predicate.

This would allow for things like counters derived on the size of
grow-only sets, booleans derived from counters, etc.

Purely a prototype; just wanted to start the conversation.