A PN-Counter for embedding in riak_dt_map
The problem with an embedded
pn-counter is when the field is removed and added again. PN-Counter
merge takes the max of P and N as the merged value. In the case
that a field was removed and re-added P and N maybe be _lower_ than
their removed values, and when merged with a replica that has not
seen the remove, the remove is lost. This counter adds some
causality by storing a `dot` with P and N. Merge takes the max
event for each actor, so newer values win over old ones. The rest
of the mechanics are the same.