Don't specify the target version in from_binary
It makes no sense to specificy a target version in `from_binary/2`
From binary should handle any binary from the past, and turn it into
the current internal representation for operation.

`to_binary/2` needs to exist so that a data type can be encoded in a way
that older modules can manipulate the data.

The problem is if the internal structure changes in such a way that
no lossless way to represent a CRDT at (say) v2 in v1.