Ensure conversions between dict/orddict-based sets and maps are total.
This addresses the issue of a top-level map with unconverted "dots" in
fields and nested CRDTs using orddict instead of dict (or vice-versa) by
recursively converting all of the necessary structures.

A callback was added to the riak_dt behavior for converting between
internal format versions.

An EQC property was added to riak_dt_map that ensures that a v2 map can
be converted completely to a v1 map, read back and up-converted to a v2
map.

NB: This does not address maps that may be partially converted to
dict; that is, nested types and fields that may still use orddict but
the top-level map using dict.

Also, riak_dt_pncounter had a notion of versions 1 and 2 that does not
correspond to this change, but we should clear up that confusion in the
future.