WIP / tentative semantic change on add
Should an add of 'X' count as a remove of all seen 'X' at the replica?
Only if there is not "action at a distance" otherwise you don't know
what's been _seen_ at this add.

This semantic is probably wrong, needs discussion/thought/work.

This change was made so that the delta matches the bigset semantics,
am I cheating?